requirements.txt
alabaster==0.7.12
argh==0.26.2
asgiref==3.2.7
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
docutils==0.16
entrypoints==0.3
flake8==3.7.9
idna==2.9
imagesize==1.2.0
Jinja2==2.11.1
livereload==2.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.3
pathtools==0.1.2
phonenumbers==8.12.1
port-for==0.3.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.6
PySocks==1.7.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.4.4
sphinx-autobuild==0.7.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
tornado==6.0.4
twilio==6.0.0
urllib3==1.25.8
watchdog==0.10.2
yapf==0.29.0
